+++ lib/server/resources/operations/host_async.go
@@ -4,28 +4,23 @@
 package operations
 
 import (
-	"fmt"
-	"time"
+	"os"
 
-	"github.com/CS-SI/SafeScale/v22/lib/utils/temporal"
 	"github.com/sirupsen/logrus"
 )
 
-func getCommand(file string) string {
-	// "sudo -b bash -c 'nohup %s > /dev/null 2>&1 &'"
-	command := fmt.Sprintf("sudo -b bash -c 'nohup %s > /dev/null 2>&1 &'", file)
-	logrus.Debugf("running '%s'", command)
-	return command
-}
+func getDefaultConnectorType() (string, error) {
+	if choice := os.Getenv("SAFESCALE_DEFAULT_SSH"); choice != "" {
+		switch choice {
+		case "cli":
+			return "cli", nil
+		case "lib":
+			return "lib", nil
+		default:
+			logrus.Debugf("unexpected SAFESCALE_DEFAULT_SSH: %s, using lib instead", choice)
+			return "lib", nil
+		}
+	}
 
-func getPhase2Timeout(timings temporal.Timings) time.Duration {
-	return timings.ContextTimeout()
-}
-
-func getPhase4Timeout(timings temporal.Timings) time.Duration {
-	return 30 * time.Second
-}
-
-func inBackground() bool {
-	return true
+	return "lib", nil
 }
